refactor : 테스트코드 도메인 영역별 패키지 분리 및 계층구조별 패키지 정리
